pyproject.toml
[project]
name = "novatrabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-sqlite3>=0.0.1",
    "openai>=1.50.0",
    "gql[all]>=3.0.0",
    "groq>=0.18.0",
    "gspread>=6.1.4",
    "oauth2client>=4.1.3",
    "psutil>=7.0.0",
    "py-cord>=2.6.1",
    "python-dotenv>=1.0.1",
]
